# server/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpsPilot API
  version: 0.7.0-dev
  description: OpsPilot auth, RBAC, Agent, Host, script, task execution, logs, scheduling, webhooks, metrics, alerts, and notifications API.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/notification-channels:
    get:
      summary: List notification channels
      description: Requires `notification:read`.
      responses:
        "200": { description: Channel list }
    post:
      summary: Create notification channel
      description: Requires `notification:write`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationChannelPayload"
      responses:
        "200": { description: Channel summary }
  /api/v1/notification-channels/{id}:
    put:
      summary: Update notification channel
      description: Requires `notification:write`. Secret fields remain write-only and are never returned.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationChannelPayload"
      responses:
        "200": { description: Channel summary }
  /api/v1/notification-channels/{id}/test:
    post:
      summary: Send test notification to channel
      description: Requires `notification:write`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: Test delivery result }
...
  /api/v1/notifications:
    get:
      summary: List notifications
      description: Requires `notification:read`.
      parameters:
        - name: unread
          in: query
          schema: { type: boolean }
      responses:
        "200": { description: Notification list }
  /api/v1/notifications/{id}/read:
    post:
      summary: Mark notification read
      description: Requires `notification:read`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: OK }
  /api/v1/audit-logs:
    get:
      summary: List audit logs
      description: Requires `audit.read`. Returns redacted audit records for the default workspace plus system-level records.
      parameters:
        - name: action
          in: query
          schema: { type: string }
        - name: actor
          in: query
          schema: { type: string }
        - name: actorType
          in: query
          schema: { type: string, enum: [user, agent, webhook, system] }
        - name: result
          in: query
          schema: { type: string, enum: [success, failed, denied] }
        - name: resourceType
          in: query
          schema: { type: string }
        - name: resourceId
          in: query
          schema: { type: string }
        - name: traceId
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/keyword"
        - name: createdFrom
          in: query
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Audit log page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"
  /api/v1/audit-logs/export:
    get:
      summary: Export audit logs
      description: Requires `audit.read`. Supports JSON and CSV export with the same filters as list.
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [json, csv], default: json }
        - name: action
          in: query
          schema: { type: string }
        - name: actor
          in: query
          schema: { type: string }
        - name: actorType
          in: query
          schema: { type: string, enum: [user, agent, webhook, system] }
        - name: result
          in: query
          schema: { type: string, enum: [success, failed, denied] }
        - name: resourceType
          in: query
          schema: { type: string }
        - name: resourceId
          in: query
          schema: { type: string }
        - name: traceId
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/keyword"
        - name: createdFrom
          in: query
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200": { description: Downloaded audit export file }
  /api/v1/audit-logs/retention/run:
    post:
      summary: Run audit log retention
      description: Requires `audit.read`. Supports dry run and execute mode.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dryRun: { type: boolean, default: true }
                days: { type: integer, minimum: 1 }
      responses:
        "200": { description: Retention execution result }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      description: Agent Token, not user JWT.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    keyword:
      name: keyword
      in: query
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    NotificationChannelPayload:
      type: object
      required: [name]
      properties:
        name: { type: string }
        channelType: { type: string, enum: [site, email, webhook, dingtalk, wechat, slack], default: site }
        config:
          type: object
          description: For `email`, use `email`; for outbound webhook-style channels, use `url` and optional `signingSecret`.
          additionalProperties: true
    AuditLogPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    AuditLog:
      type: object
      properties:
        id: { type: integer }
        actorType: { type: string, enum: [user, agent, webhook, system] }
        actorUser: { type: string }
        actorAgent: { type: string }
        action: { type: string, example: task.create }
        resourceType: { type: string, example: task_run }
        resourceId: { type: integer }
        result: { type: string, enum: [success, failed, denied] }
        ip: { type: string }
        userAgent: { type: string }
        traceId: { type: string }
        requestMethod: { type: string, example: POST }
        requestPath: { type: string, example: /api/v1/tasks }
        before: { type: string, description: Redacted JSON string. }
        after: { type: string, description: Redacted JSON string. }
        metadata: { type: string, description: Redacted JSON string. }
        createdAt: { type: string }